feat(infra/azure-functidons): add infra config iteration for apply and plan mode

+++ infra/azure-functions/src/azure-command.ts
@@ -1,6 +1,6 @@
 import { $ } from 'bun';
 
-import type { InfraEnvConfig } from './types';
+import type { InfraEnvConfig, InfraOptions } from './types';
 
 import { fallbackResourceIdentifier, toTarget } from './utils';
 
@@ -21,9 +21,9 @@ export function getResourceName(infraConfig: InfraEnvConfig): { prefix: string; 
   };
 }
 
-export async function createFunctionApp(infraConfig: InfraEnvConfig) {
+export async function createFunctionApp(infraConfig: InfraEnvConfig, options: InfraOptions) {
   const resourceName = getResourceName(infraConfig);
-  console.log('Creating function app');
+  console.log('-'.repeat(80));
 
   const location = 'eastasia';
   const resourceGroup = `rg-nammatham-${resourceName.prefix}`;
@@ -34,20 +34,11 @@ export async function createFunctionApp(infraConfig: InfraEnvConfig) {
   }
   const osType = infraConfig.platform === 'linux' ? 'Linux' : 'Windows';
 
-  //   # Create a resource group
-  // echo "Creating $resourceGroup in "$location"..."
-  // az group create --name $resourceGroup --location "$location" --tags $tag
-
-  // # Create an Azure storage account in the resource group.
-  // echo "Creating $AZURE_STORAGE_ACCOUNT"
-  // az storage account create --name $AZURE_STORAGE_ACCOUNT --location "$location" --resource-group $resourceGroup --sku $skuStorage
-
-  // # Set the storage account key as an environment variable.
-  // # export AZURE_STORAGE_KEY=$(az storage account keys list -g $resourceGroup -n $AZURE_STORAGE_ACCOUNT --query '[0].value' -o tsv)
-
-  // # Create a serverless function app in the resource group.
-  // echo "Creating $functionApp"
-  // az functionapp create --name $functionApp --storage-account $AZURE_STORAGE_ACCOUNT --consumption-plan-location "$location" --resource-group $resourceGroup --os-type Linux --runtime custom --functions-version $functionsVersion
+  if (options.isPlanMode) {
+    console.log('Plan mode enabled, skipping resource creation');
+    return;
+  }
+  console.log('Creating function app');
 
   console.log(`Creating ${resourceGroup} in ${location}...`);
   await $`az group create --name ${resourceGroup} --location ${location}`;
@@ -58,15 +49,26 @@ export async function createFunctionApp(infraConfig: InfraEnvConfig) {
   console.log(`Creating function app: ${resourceName.prefix}`);
   await $`az functionapp create --name ${resourceName.prefix} --storage-account ${resourceName.shortPrefix} --consumption-plan-location ${location} --resource-group ${resourceGroup} --os-type ${osType} --runtime custom --functions-version ${functionsVersion}`;
 
+  console.log(`Set Function Key for ${resourceName.prefix} `);
+  await $`az functionapp keys set --name ${resourceName.prefix} --resource-group ${resourceGroup} --key-name github_actions --key-value ${process.env.AZURE_FUNCTIONS_API_KEY}`;
+
   console.log('Function app created');
+  console.log('-'.repeat(80));
 }
 
-export async function destroyFunctionApp(infraConfig: InfraEnvConfig) {
+export async function destroyFunctionApp(infraConfig: InfraEnvConfig, options: InfraOptions) {
+  console.log('-'.repeat(80));
   const resourceName = getResourceName(infraConfig);
+
+  if (options.isPlanMode) {
+    console.log('Plan mode enabled, skipping resource creation');
+    return;
+  }
   console.log('Destroying function app');
 
   const resourceGroup = `rg-nammatham-${resourceName.prefix}`;
 
   console.log(`Deleting ${resourceGroup}`);
   await $`az group delete --name ${resourceGroup} --yes`;
+  console.log('-'.repeat(80));
 }
